# internal/web/apidoc/docs/swagger.yaml
basePath: /
definitions:
  apidoc.JsonResponseError:
    properties:
      code:
        example: 0
        type: integer
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        example: Validation failed
        type: string
      success:
        example: false
        type: boolean
    type: object
  apidoc.JsonResponseException:
    properties:
      code:
        example: 0
        type: integer
      exceptionType:
        type: string
      message:
        example: Internal Server Error
        type: string
      success:
        example: false
        type: boolean
    type: object
  apidoc.JsonResponseOk:
    properties:
      data: {}
      message:
        example: ""
        type: string
      success:
        example: true
        type: boolean
    type: object
  apidoc.JsonResponseUnauthorized:
    properties:
      code:
        example: 0
        type: integer
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        example: Access token not found
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  model.AccountResult:
    properties:
      balance:
        type: string
      currency:
        $ref: '#/definitions/model.CurrencyResult'
      folderId:
        type: string
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      owner:
        $ref: '#/definitions/model.UserResult'
      position:
        type: integer
      sharedAccess:
        items:
          $ref: '#/definitions/model.SharedAccess'
        type: array
      type:
        type: integer
    type: object
...
  model.CreateRecurringTransactionRequest:
    properties:
      accountId:
        type: string
      accountRecipientId:
        type: string
      amount:
        $ref: '#/definitions/vo.FlexString'
      categoryId:
        type: string
      description:
        type: string
      id:
        type: string
      nextPaymentAt:
        type: string
      payeeId:
        type: string
      schedule:
        type: string
      tagId:
        type: string
      type:
        type: string
    type: object
  model.CreateRecurringTransactionResult:
    properties:
      item:
        $ref: '#/definitions/model.RecurringTransactionResult'
    type: object
...
  model.CurrencyResult:
    properties:
      code:
        type: string
      fractionDigits:
        type: integer
      id:
        type: string
      name:
        type: string
      symbol:
        type: string
    type: object
  model.CurrentUserResult:
    properties:
      accessLevel:
        type: string
      accessUntil:
        type: string
      avatar:
        type: string
      currency:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/model.OptionResult'
        type: array
      reportPeriod:
        type: string
    type: object
...
  model.DeleteRecurringTransactionRequest:
    properties:
      id:
        type: string
    type: object
  model.DeleteRecurringTransactionResult:
    type: object
...
  model.GetRecurringTransactionListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/model.RecurringTransactionResult'
        type: array
    type: object
...
  model.OptionResult:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
...
  model.PostRecurringTransactionRequest:
    properties:
      accountId:
        type: string
      accountRecipientId:
        type: string
      amount:
        $ref: '#/definitions/vo.FlexString'
      amountRecipient:
        $ref: '#/definitions/vo.FlexString'
      categoryId:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      payeeId:
        type: string
      recurringId:
        type: string
      tagId:
        type: string
      type:
        type: string
    type: object
  model.PostRecurringTransactionResult:
    properties:
      accounts:
        items:
          $ref: '#/definitions/model.AccountResult'
        type: array
      item:
        $ref: '#/definitions/model.TransactionResult'
      nextPaymentAt:
        type: string
    type: object
  model.RecurringTransactionResult:
    properties:
      accountId:
        type: string
      accountRecipientId:
        type: string
      amount:
        type: string
      categoryId:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      nextPaymentAt:
        type: string
      ownerUserId:
        type: string
      payeeId:
        type: string
      schedule:
        type: string
      tagId:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
...
  model.SharedAccess:
    properties:
      isAccepted:
        type: integer
      role:
        type: string
      user:
        $ref: '#/definitions/model.UserResult'
    type: object
...
  model.SkipRecurringTransactionRequest:
    properties:
      id:
        type: string
    type: object
  model.SkipRecurringTransactionResult:
    properties:
      item:
        $ref: '#/definitions/model.RecurringTransactionResult'
    type: object
...
  model.TransactionResult:
    properties:
      accountId:
        type: string
      accountRecipientId:
        type: string
      amount:
        type: string
      amountRecipient:
        type: string
      author:
        $ref: '#/definitions/model.UserResult'
      categoryId:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      payeeId:
        type: string
      tagId:
        type: string
      type:
        type: string
    type: object
...
  model.UpdateActiveBudgetRequest:
    properties:
      value:
        type: string
    type: object
  model.UpdateActiveBudgetResult:
    properties:
      user:
        $ref: '#/definitions/model.CurrentUserResult'
    type: object
...
  model.UpdateRecurringTransactionRequest:
    properties:
      accountId:
        type: string
      accountRecipientId:
        type: string
      amount:
        $ref: '#/definitions/vo.FlexString'
      categoryId:
        type: string
      description:
        type: string
      id:
        type: string
      nextPaymentAt:
        type: string
      payeeId:
        type: string
      schedule:
        type: string
      tagId:
        type: string
      type:
        type: string
    type: object
  model.UpdateRecurringTransactionResult:
    properties:
      item:
        $ref: '#/definitions/model.RecurringTransactionResult'
    type: object
...
  model.UserResult:
    properties:
      avatar:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  vo.FlexString:
    type: object
info:
  contact: {}
  description: Self-hosted personal finance and budgeting API
  title: Econumo API
  version: 1.0.0
paths:
...
  /api/v1/recurring/create-recurring-transaction:
    post:
      consumes:
      - application/json
      description: Creates a recurring transaction template. Idempotent on the request
        id.
      parameters:
      - description: Create recurring transaction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateRecurringTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.CreateRecurringTransactionResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: Create a recurring transaction
      tags:
      - Recurring
  /api/v1/recurring/delete-recurring-transaction:
    post:
      consumes:
      - application/json
      description: Deletes an existing recurring transaction template.
      parameters:
      - description: Delete recurring transaction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DeleteRecurringTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.DeleteRecurringTransactionResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: Delete a recurring transaction
      tags:
      - Recurring
  /api/v1/recurring/get-recurring-transaction-list:
    get:
      description: Returns every recurring transaction template on accounts the caller
        can access.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.GetRecurringTransactionListResult'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: List recurring transactions
      tags:
      - Recurring
  /api/v1/recurring/post-recurring-transaction:
    post:
      consumes:
      - application/json
      description: Materializes one instance of a recurring transaction template as
        a real transaction and advances its schedule. Idempotent on the request id.
      parameters:
      - description: Post recurring transaction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PostRecurringTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.PostRecurringTransactionResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: Post a recurring transaction
      tags:
      - Recurring
  /api/v1/recurring/skip-recurring-transaction:
    post:
      consumes:
      - application/json
      description: Advances a recurring transaction template's schedule without posting
        a transaction.
      parameters:
      - description: Skip recurring transaction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SkipRecurringTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.SkipRecurringTransactionResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: Skip a recurring transaction
      tags:
      - Recurring
  /api/v1/recurring/update-recurring-transaction:
    post:
      consumes:
      - application/json
      description: Updates an existing recurring transaction template.
      parameters:
      - description: Update recurring transaction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateRecurringTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.UpdateRecurringTransactionResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: Update a recurring transaction
      tags:
      - Recurring
...
  /api/v1/transaction/export-transaction-list:
    get:
      description: Returns a CSV (text/csv) of transactions on the selected accounts
        (comma-separated accountId), or all accessible accounts when omitted.
      parameters:
      - description: Comma-separated account ids
        in: query
        name: accountId
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: Export the transaction list as CSV
      tags:
      - Transaction
...
  /api/v1/user/update-budget:
    post:
      consumes:
      - application/json
      description: Sets the authenticated user's default budget and returns the refreshed
        user.
      parameters:
      - description: Update budget request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateActiveBudgetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.UpdateActiveBudgetResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: Update budget
      tags:
      - User
...
securityDefinitions:
  Bearer:
    description: 'Opaque access token. Format: "Bearer <token>".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"